refactor(PR-667): generators
